Coming soon update...
In this commit, the site now has a static initial page that provides
the company’s logo and name, with an alert message that is displayed
upon the beginning of a session. Also, this alert is displayed via
click event, triggered by the user, while the user is clicking the help
button.